+++ examples/webpack/src/app/pages/lazy/lazy.component.ts
@@ -279,12 +279,14 @@ export class LazyComponent {
     style: {
       bgColor: 'rgba(255,0,0,.5)',
       textColor: 'blue',
-      margin: {
-        marginTop: '3px',
-        marginBottom: '0px',
-        marginLeft: '5px',
-        marginRight: '5px'
-      }
+      marginTop: '3px',
+      marginBottom: '0px',
+      marginLeft: '5px',
+      marginRight: '5px',
+      position: 'absolute',
+      top: '0px',
+      height: '25px'
+      // be careful to use width, in particular with % values
     }
   };